Use fail/then instead of error/done From the jQuery upgrade guide:
The jqXHR object returned from jQuery.ajax() is a jQuery Deferred and
has historically had three extra methods with names matching the
arguments object of success, error, and complete. This often confused
people who did not realize that the returned object should be treated
like a Deferred. As of jQuery 3.0 these methods have been removed.
As replacements, use the Deferred standard methods of done, fail,
and always, or use the new then and catch methods for Promises/A+
compliance.

Note that this does not have any impact at all on the ajax callbacks
of the same name passed through the options object, which continue to
exist and are not deprecated. This only affects the jqXHR methods.