Move frontend installation recipes to local app templates
This change makes it trivial to add/remove/change frontend or to pass
a custom template to be used. This among other use-cases can become
handy for testing out branches of solidus_starter_frontend without
touching the code of the installer.

It's also closer to what you would expect coming from `rails new`
options.

In the process `--frontend=none` was fixed and now properly removes
`solidus_frontend`.

The bundler splitter was compacted into a single file and always adds
`solidus_frontend` leaving the choice to other templates if to keep it.