Do not discard prices upon variant deletion
When we soft-delete a variant, we also soft-delete the variant's prices.
In my opinion, this is not necessary, as the association between prices
and variants includes discarded prices, and harmful, as it leads to
price inconsistencies between a variant before discarding and after
undiscarding (if the variant has a newer, but discarded, price, we have
no way of knowing that that price should not be un-discarded along with
the variant).

(cherry picked from commit 5d8531d5076e4094338c6c32a992005f2094874e)