Call the solidus generator by shelling-out for the dummy app
We already shell-out for setting the environment and running
migrations, this way we can avoid requiring the generator and we run
it in a standard app environment instead of from within a library.

Before this change it was failing to set Rails.root as the app_path
for the generator.