Maintain prices for previously soft-deleted variants
Previous Solidus behavior included soft-deleted prices when looking at
the default price. The intention of this behavior was that a a variant
that is soft-deleted, and has its prices soft-deleted as a consequence,
still has a `default_price` record.

However, for a `kept` variant, we do not want to see a discarded
`default_price` - otherwise what sense does it make to even add the
delete action to the price? Especially given that deletion touches the
variant and will then move it forward in the array of sorted prices to
be considered for a given set of pricing attributes.

This commit changes the `prices` association to include discarded
prices, but then filters out discarded prices for non-discarded
variants in the price selector.