Remove empty country_names translations
The way this is used, these empty translations will break portions of
the admin, because the root "country_names" key is accessed and used as
a hash. This breaks the admin in certain places if you use one of these
locales, because it will use nil as the name for the countries and then
try to call string methods on it.