Add support for sorting store credits with different algorithms
In our store, we have a set of custom eligibility rules for store
credits based on the contents of the order. So we need a hook to
inject a different way to prioritize store credits, as opposed to
simply the type priority. While it might be possible to monkey patch
the `in_priority_order` scope, that feels somewhat wrong - and the
`add_store_credit_payment` method of `Spree::Order` does a *lot*
that we would prefer not to monkey patch.

This PR adds a very simple configuration hook in the middle of this
method to allow stores to inject some new behavior. By default, this
simply does what the existing method did - sort by credit priority.