GitHub Action to automatically create backport PRs
This commit integrates the backport-github-action [1] GitHub Action.

It automatically creates PRs for the branches specified via labels on
another PR. For instance, if a PR to master has the following labels:

- Needs Backport
- Needs Backport-v3.2
- Needs Backport-v3.1

It'll create two other PRs, cherry-picking the commits on v3.2 and v3.1
branches. The action is triggered whenever:

- A PR with those labels is merged
- A closed PR has those labels added

After the action runs, the "parent" PR is updated with a status report,
informing of any encountered error (e.g., merge conflicts).

We've chosen this action over the Backporting [2] action because the
former has support for PRs containing multiple commits, which
corresponds to our current flow. At the same time, the latter only works
with single-commit PRs.

1 - https://github.com/sqren/backport-github-action
2 - https://github.com/marketplace/actions/backporting