Assign option types to all clothing samples
It is unexpected that the loading of samples only provides size and
colour options to Solidus T-Shirt. This commit creates
`Spree::ProductOptionType`s for all clothing options and adds size
variants for the clothes which were missing any.

It also updates the names of the generated `Spree::OptionType`s to match
the ones from demo.solidus.io.