Automatically update the release draft when merging PRs
We add a GitHub action to automatically update the next release draft
[1] whenever a PR is merged.

A PR is categorized under at least one of the changelog categories
according to its labels [2]. Labels are already automatically added by the
labeler GitHub action [3]. A PR can be skipped when the `changelog:skip`
label is added.

The chosen format mimics what we've done in the Changelog [4], so we can
leverage this work further to automate its generation.

We prepend the draft content with a comment warning about its automatic
generation and append a link to the list of new commits since the
previous tag.

The candidate tag is computed like this:

- We take the current tag in the base branch thanks to a custom callable
  workflow that extracts the information from the pipeline.
- For PRs to master, the following tag is the subsequent minor or major
  version depending on the value of `Spree::VERSION`.
- The following tag is the next patch version for PRs to branches
  corresponding to supported versions.

We considered using Release Drafter action [5] as an alternative, but it
reaches GitHub API abuse limit in our repo.

[1] - https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository
[2] - https://github.com/solidusio/solidus/pull/4756
[3] - https://github.com/actions/labeler
[4] - https://github.com/solidusio/solidus/blob/0cae150e4f347cd9f232001bb945ecc9974abb29/CHANGELOG.md
[5] - https://github.com/marketplace/actions/release-drafter