Revert jQuery Changes to Xhr Var in Image Upload
The was changed by mistake to `$.ajaxSetup.xhr()` and broke the image
upload drop field. The error was missed because there were not any
existing tests for the image drop field so the script was not being
properly tested.

Although `$.ajaxSetup().xhr()` could work in this situation, it will
be fully reverted to original due to inability to find source of
deprecation cited in commit that changed this function.