Allow productAutocomplete to support a custom ransack query
From this change, when initializing a productAutocomplete, a function
can be given to the searchCallback option argument. This callback should
return an object (hash), and accepts a parameter which is the currently
typed string in the autocomplete select2. The "q" key on the returned
hash will be given to Ransack.