Require mail >= 2.8.0.rc1 for Ruby 3.1 support
Ruby 3.1 removed net-smtp, net-imap and net-pop from
standardlib and extracted them into independent gems.

The mail gem >= 2.8.0.rc1 added them to their dependencies.
Since we require the mail gem in core.rb:17 we need to
add the mail gem as dependency and need to make sure
it is the Ruby 3.1 compatible version as we support
Ruby 3.1