Make split_promotions_with_any_match_policy task faster
Do not re-apply promotions to incomplete orders if there
are no promotions effected. This is not necessary and speeds
up the task for shop with lots of incomplete orders (this could
easily be in the thousands).

Also make the effected promotions update faster by using find_each
and include actions as well.