Ignore big reformatting commits in GitHub blame UI
By adding a .git-blame-ignore-revs file that contains the commits that
git blame will ignore, we are able to remove them from the blame view
when using this feature.

See https://docs.github.com/en/repositories/working-with-files/using-files/viewing-a-file#ignore-commits-in-the-blame-view
for extra details.

For now, I added just a couple of big commits, but we can add more with
time.

This same behavior can be enabled locally with:

	git config blame.ignoreRevsFile .git-blame-ignore-revs

This config is not suitable as a global config though, because it will
fail on repositories that do not have the file set.