Improve variantAutocomplete to be more reusable
From this change, when initializing a variantAutocomplete, a function
can be given to the searchCallback option argument.

Before, variantAutocomplete accepted a object (hash) searchOptions
argument, now, instead, searchCallback should be used and given a
function that returns an object (hash). It accepts a parameter which is
the currently typed string in the autocomplete select2. The "q" key on
the returned hash will be given to Ransack if the variant_search_term
key is overridden with a falsy value.

For compatibility, if searchOptions was being used, searchCallback can
just return the same hash.