Use labeler action to mark all kind of changes
We leverage the labeler action [1] to mark all kinds of changes done in
the repository. We will reach the end goal with future work that will
comprise:

- Ensuring that at least one changelog label is set before merging a PR.
- Leveraging the changelog labels to build both the GH release draft and
  the Changelog automatically.

The expected labels are:

- `changelog:solidus_core` for changes relative to the `solidus_core`
  gem.
- `changelog:solidus_backend` for changes relative to the
  `solidus_backend` gem.
- `changelog:solidus_api` for changes relative to the `solidus_api` gem.
- `changelog:solidus_sample` for changes relative to the
  `solidus_sample` gem.
- `changelog:solidus` for changes relative to the `solidus` gem.
- `changelog:repository` for changes done to code that doesn't belong to
  any gem.

We'll also manually add a `changelog:skip` label that we will use to
override the automatic inclusion of a change. We'll also probably skip
entries relative to `changelog:repository`.

References #4752

1 - https://github.com/actions/labeler