Remove position column from spree_taxons
This column has not been used since 8d7d4d7, which had been merged into
Spree 1.1.0 in 2013.

Background: Apparently, taxons used to be a list in very early versions
of Spree, and have then been refactored to be a nested set. Nested sets
are strict supersets of lists, so this column is entirely unnecessary.