Ensure the presence of a changelog label before merging a PR
With #4756, we created a set of labels prefixed with `changelog:`. They
cover all possible changes in the repository. We're now adding a check
to ensure that at least one is present.

The action runs in two contexts:

- When labels on a PR change.
- Just after the automatic labeler has done its work, which is a
  safeguard in case we add new files to the repo that don't meet the
  current rules.

We need to query the API with the `gh` command as `${{
github.event.pull_request.labels }}` won't contain the updated
information after the `triage` step on the "Triage for Changelog"
workflow. For simplicity, we keep the same for "Ensure Changelog label"
standalone workflow.

Closes #4765