Add action to prepare a release
We add a GitHub action that automates the chores to prepare a new
Solidus release. It creates a PR that:

- It bumps the version in `core/lib/spree/core/version.rb` [1].
- It bumps the version in `docker-compose.yml` [2]
- It updates the Changelog with the contents extracted from the
  automatically generated Release Draft on GitHub [3].

As the contents of the Changelog will be automatically generated from
now on, we're removing the manual additions we had done for the
next-to-be version. We'll be moving those instructions to a dedicated
page on our documentation site [4], providing a better experience
altogether.

The added GitHub action needs to be manually dispatched [5]. We can do it
either through GitHub's UI (Actions tab) or the GitHub API. For the
latter, GitHub's CLI [6] comes as handy to make it more concise:

gh workflow run "Prepare release"

It's worth noticing that, for minor and major releases, we'll be
creating the last commit for a release in the master branch instead of a
new patch branch. On https://github.com/solidusio/solidus/pull/4719,
we'll automatically generate the patch branch from that commit so the
tag will be present in both branches. We're also adding the release date
on the actual release that will happen right after, while until now, we
were updating that information only on the master branch.

Solidus maintainers will still have to run the command `gem release`
themselves once the PR is merged. Even if we could also automate that,
RubyGems would currently require us to give up MFA [7], and we don't want to
compromise security.

[1] - https://github.com/solidusio/solidus/blob/cc214550222561fd0ffdce334f1dd1e1637354d8/core/lib/spree/core/version.rb
[2] - https://github.com/solidusio/solidus/blob/cc214550222561fd0ffdce334f1dd1e1637354d8/docker-compose.yml
[3] - https://github.com/solidusio/solidus/pull/4719
[4] - https://guides.solidus.io/
[5] - https://docs.github.com/en/actions/managing-workflow-runs/manually-running-a-workflow
[6] - https://cli.github.com/
[7] - https://guides.rubygems.org/setting-up-multifactor-authentication/