Automate chores after releasing a new version
After automating the setup for a new release in #4791 [1], this PR adds
another GitHub action to streamline the chores that need to be done
after that.

We're adding the "release-generate" label to the PR generated after
setting up for a release. When merged, PRs with that label will take
care of triggering the new automation:

- Bumping the just-released version number to the `.dev` stage of the next
one.
- Bumping the versions on `docker-compose.yml` and
  `Spree.previous_solidus_minor_version` when the base branch is master.
- Publishing the matching release notes and creating the corresponding
  tag.
- Creating a new branch when the base branch is master.

When we set up for a new release on master, we need to have in mind
whether the next one after that will be another minor or a major. That
will change the development version we need to bump to. Because of that,
we're adding a new boolean parameter to the manually dispatchable action
to set up the release. When we specify that the next one is going to be
a major, we also add the `release-generate-last_minor` label to the new
PR and use it to calculate different version targets.

We're updating our conventions so that all branches point to the `.dev`
version of the next release. Only the last commits to be included in the
releases got that suffix removed.

[1] - https://github.com/solidusio/solidus/pull/4791