Improve Taxon factory
When creating a Taxon under a Taxonomy normally via admin it will
make set the taxonomy correctly and the parent to a given specific
parent or the the taxonomy root by default.

This makes the factory behave in a similar way.

If you want a root Taxon, you should create a Taxonomy and grab the
root, as a root is automatically created upon creating a Taxonomy.

Note, some reason icons were not being correctly set by the factory for
nested Taxons so it had to be changed to do after creation.