Allow only one root Taxon per Taxonomy
Spree::Taxonomy rightfully assumes there's only one root [1]. Before,
it was it easy to create another "root" taxon, which would not appear
in the admin and/or the Taxonomy would return the wrong root. This is
because `parent_id: nil` is the defining factor of a root.

As the validation is checking uniqueness on taxonomy_id, a spec was
added to ensure a Taxonomy can still have multiple Taxons (just not
multiple ones with a parent_id of nil).

[1] https://github.com/solidusio/solidus/blob/2b79f72aa53f5caa850c587888fff46c1c91f7b7/core/app/models/spree/taxonomy.rb#L10