Fix source tag used to create the diff for the release notes on master
On the automatic generation of the release notes (see #4719), we append
them with a link to the diff of commits after the new release. E.g.:

**Full Changelog**: https://github.com/solidusio/solidus/compare/v3.2.1...v3.2.2

Using the current tag in the series as the diff source works for
patch-level releases. However, in minor and major releases, we need to
use the first patch level of the previous version (e.g., v3.2.0) for a
couple of reasons:

- That's what it really will show all the changes done in master since the
last release.
- Otherwise, whenever a new patch version is released, the automatic
generation of the release notes will break, as the expected appended
will have changed.

Fixes #4795