Be explicit about the solidus_frontend gemspec dependency
We need to avoid being too strict on the frontend dependency
because otherwise we need to release a new solidus_frontend
.dev version everytime we release a new solidus 3.2.x patch
level version.

With this change, whenever the solidus main gem is used, Bundler
will compute the frontend dependency less strictly and will pick
the last version released on RubyGems.