Add a default implementation for PaymentMethod#try_void
We use to replicate this behavior on all the payment extensions
that add a payment integration. Still, the logic is pretty similar
all the times.

By adding this default behavior we can save some time integrating
new payment methods, unelss something special is needed.

The default try_void method added will just check if the void attempt
is successful, otherwise it will return false. This will determine if
Solidus will emit a refund along with an order cancellation.

The method is a bit complex because void can have different arity, based
on if the payment method support payment profiles or not. See [1] for
more information.

One thing worth mentioning: the self.name definition in the anonymous
payment method classes in tests is needed for Rails 5.2.

[1]: https://github.com/solidusio/solidus/blob/16d1e3704fbbe630a584c919de1a640a6ab6c6f8/core/app/models/spree/payment/processing.rb#L74-L81