Raise a custom extension passing invalid search params
Right now, passing a string or any other thing that is not
a Hash to the searcher in the `:search` parameter will raise
a NoMethodError.

For example: `/products?per_page=&keywords=&search=blub` would raise
a NoMethodError.

This commit avoids that a generic NoMethodError is raised and instead
raise a very specific error that serves two purposes:

1. Helps debugging: the error message contains the name of the param
   that is invalid, making it easier to spot eventual issues.
2. Allows to rescue the error safely at the storefront level. For
   example, to respond with a 400, one can easily do:

   rescue_from Spree::Core::Search::Base::InvalidOptions do |error|
     raise ActionController::BadRequest.new, error.message
   end