Ensure Taxonomy names are unique
In the next commit, a validation will be added to ensure their names
are unique if they belong to the same parent. This will indirectly
cause Taxonomies roots to fail if the Taxonomy name already exists.
This will mean the Taxonomy will be created but not the root. To
prevent creation of unuseful Taxonomy, a validation was added to
prevent this at a higher level.