Give allowed_ransackable_* methods a default value
Having a default of `[]` seems more appropriate than `nil`. This will
also allow stores to concatenate onto the default value and get
automatic future additions in Solidus, whereas before, stores would
have to define the array in their decorator if the core model did not set any values.

This also applies to the deprecated whitelisted_ransackable_* as it
delegates to allowed_ransackable_*.