Remove Simple/BogusCreditCard#try_void
It will fall back to Spree::PaymentMethod#try_void
via class inheritance.

Now that the parent class' try_void has an implementation,
we can rely on that for the Simple/BogusCreditCard.

Before this change, all the attempts to voiding a payment in
the admin panel would be succeding. Now we can simulate a
failure (by trying to cancel an order with captured payment)
and this will create a refund for that payment instead.

Some specs have been updated as well:

- order state machine: refactored and added some scenarios.
  Now the specs are only testing that cancel! on payment is
  actually called, leaving to Spree::Payment::Cancellation
  and its specs the responsibility to test what happens when
  it's called.
- payment method: to simulate what happens with real payment
  methods, we used to create an anonymus class with the expected
  behavior. This behavior is now identical to the default
  implementation, used by simple bogus credit card and bogus
  credit card. We can just use the factories.