Remove duplicated Order#cancel! tests
We had a lot of duplication between specs in order_spec
and order/state_machine_spec.

I tried to make some order, moving all the cancel! logic
specs into order_spec and leaving to order/state_machine
the only responsibily to test the value of can_cancel?
because that's the only thing that is done by the state
machine, other than calling Spree::Order#after_cancel
when the transition happens.