Fetch solidus_frontend from RubyGems instead of GitHub
Fixes https://github.com/solidusio/solidus/issues/4878.

With this change, we're breaking from the practice of testing Solidus
against the edge version of SolidusFrontend. We're assuming that there
won't be significant development on SolidusFrontend and thus it would be
safe to test Solidus against the latest stable version of
SolidusFrontend.