Fix CI only testing with the legacy event system adapter
From https://github.com/solidusio/solidus/pull/4666 all CI jobs have
been running with the legacy event system. We were defaulting the
`legacy` parameter to `0`, which we copied over to the
`USE_LEGACY_EVENTS` environment variable. However, the test suite was
checking for the mere existence of the environment variable to select
the legacy adapter and not for an actual `1` value.

We are now:

- Renaming the `legacy` parameter to `legacy_events` to be clearer.
- Making the parameter a boolean one, defaulting to false.
- Checking that `USE_LEGACY_EVENTS` is the string `"true"` before
  selecting the legacy adapter.
- Using the legacy adapter for the job on Ruby 2.7 & Rails 6.0.
- Adding `-legacy_events` to the job name when selected.