Fix sticky admin nav on long menu
In Solidus, the admin nav menu becomes sticky after scrolling the page.
If the menu contains a few entries, everything works fine.
But if the menu includes some entries and the height exceed the window
height, since the menu is sticky, the latest entries will be hidden.

This commit applies the sticky only if the menu height does not exceed
the window height.