Fix CI only testing with the paperclip adapter
Since https://github.com/solidusio/solidus/pull/4666, all jobs have been
running with the paperclip adapter.

The reason is that when the `paperclip` parameter was `false` on
CircleCI, that translated into the `DISABLE_ACTIVE_STORAGE` env var being
the string `"false"`. As we were checking for the mere presence of
`DISABLE_ACTIVE_STORAGE`, it always evaluated to `true`.

Fixes #4901