Fix links to PR & author not being added into the automatic Changelog
Although GitHub automatically autolinks references to PR and author in
the release notes, it doesn't create them for the files in a repository.
From theirs docs [1]:

> Note: Autolinked references are not created in wikis or files in a repository.

As we're copying the automatically generated release notes into the
Changelog file, that resulted in the missing links. We fix it from the
source to be safe from future changes.

[1] - https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/autolinked-references-and-url

Closes #4904