Allow a payment id to be provided during checkout
This allows modern payment methods (relying on tokens and frontend
confirmation) to have a record on solidus' side that can track events
and the payment status.

Creating a payment before or during the payment step of the checkout
was already possible, but resulted in an additional payment to be
created before transitioning to the confirm step and invalidating all
previous payments.

This way it can simply be provided via its id, and fetched from the
payments currently associated with the order.

If payment methods will handle their own payment creation in the
future we can distance ourselves from the credit-card coupled code and
have a generic payment interface that can better serve modern payment
methods.