Let the sandbox use javascript
New Rails versions do not come with webpacker anymore.
We can safely generate sandboxes with Javascript now.

This is also needed because now the Starter Frontend has
some Stimulus/Turbo code that relies on the Rails JS stack
to work.

This change needs to bump the CircleCI cache for the installer
test, because the Rails application needs to be re-installed
without the --skip-javascript option.