Use the same format as GitHub's default for the release notes
We're changing the format for the auto-generated release notes from:

- This is the PR title #1 (@author)

to:

* This is the PR title by @author in https://github.com/solidusio/solidus/pull/1

By mimicking the default GitHub format, we make it easier to generate
them manually through GitHub's "Generate release notes" [1] button or
through the API [2].

At the same time, we're fixing the missing links to the PR when we copy
the draft content to the Changelog file, as GitHub doesn't autolink
references in files. From their documentation [3]:

> Note: Autolinked references are not created in wikis or files in a repository.

We'll still miss the link to the contributor, but it's a tradeoff we pay
to make our lives easier if the automation fails. Anyway, it can be
accessed once landed on the pull request URL.

[1] - https://docs.github.com/en/repositories/releasing-projects-on-github/automatically-generated-release-notes
[2] - https://docs.github.com/en/rest/releases/releases?apiVersion=2022-11-28#generate-release-notes-content-for-a-release
[3] - https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/autolinked-references-and-url

Closes #4904 & #4908