Prevent Taxon factory spec false positive
On a clean database, the spec was creating a Taxonomy and its root
Taxon with an id of 1. This means it does not accurately test
the subject Taxon's parent_id (setting the parent_id as the Taxonomy
id would assert truthy).

To prevent this, a Taxon is created before the spec, so the Taxonomy
root Taxon has a different id to the Taxonomy (2 and 1 respectively).