Copy new migrations as part of the update task
Right now, we are not copying the new migrations
introduced with the new version. This new install
step will copy migrations and it's left to the
user to run db:migrate, before checking them.

This commit also update some copy of the task to
reflect the new content of the file.