Add back Variant#find_or_build_default_price
This method was renamed to `default_price_or_build` with 6420b10, but
this was a breaking change given it was released with Solidus 3.1.
This commit reintroduces the method as an alias but with deprecation.

(cherry picked from commit 65a6f2d5dfb3e8af4bab2d649149fb36ea42d93a)