Update descend_by_popularity scope spec
so that it doesn't depend on the line item quantity.

The scope is not taking into account the line item quantity
and was actually evaluating both line items with the same
order. variant_1 was picked first only because it was
created first.

In fact, switching the creating order of the two variants
it was failing consistently.

The flaky was due to the fact that it's not guarantee that
the databases will return the record in the order they are
created, unless specified.