Fix rake error testing the update generator
It looks like the `spree:install:migrations` task is not available when
running through `Rails::Generators.invoke`, as it context differs from
common shell invocation.

We fix that by not trying to run the task on test environment. In the
end, it'd be a good idea to copy the migrations at the test suite
runtime.