Remove support for deprecated promo rules matching policy
Matching "all" rules is the only supported path now. You can have the
same effect of the previoys "any" matching policy by creating multiple
promotions with the same configuration.

This change also removes the rake task we provided to update the
existing promotions. That script needs to be executed on previous
versions of Solidus only.

Ref https://github.com/solidusio/solidus/pull/4304