Deprecate Spree::Promotion::MATCH_POLICIES without replacement
This constant won't be needed anymore in 4.0 when we will remove
the any matching policy for promotion rules.

I tried to use a more simple ActiveSupport::Deprecation::DeprecatedObjectProxy
like:

    MATCH_POLICIES = ActiveSupport::Deprecation::DeprecatedObjectProxy.new(
      %w(all any),
      "Spree::Promotion::MATCH_POLICIES is deprecated, please do not use it anymore.",
      Spree::Deprecation
    )

But it prints the deprecation warning twice for some reason. The
approach used in this commit will load the file that contains the
constant only when it's called, which will also print the warning.