Add Spree::Migration
Uses the current Rails version the app is running to run migrations.
This makes sure, that a) extensions that are running an older Rails
version (we still support down to Rails 5.2) are able to run migrations
that are introduced in later Rails versions. Also it makes sure that
the Solidus tables are created the same way as the rest of the host
apps tables that most likely exist. Most of our migrations or very old
(some of them still run with Rails 4.2 version). This changes tables
and timestamps to very old Rails defaults. Leading to a misconfigured
and inconsistent database schema for new Rails apps. Since this does
not effect existing apps, it should be safe to do that.