Do not install paypal commerce plattform on auto-accept
Whenever we run the installer in ie. a dummy app for an extension
on a CI server with --auto-accept option we now get the paypal commerce
plattform extension installed. Most extension tests are very fine
with the dummy payment methods core provides and does not need
a real payment method (that wont work in a dummy env anyway).

If we want to install this payment method for new stores or the
demo should explicitely do that and not without asking.