Add webrick as explicit dev dependency
We need webrick to run JS tests using Teaspoon.
This dependency is no more bundled with latest ruby versions
and won't be available unless explicitely added to the Gemfile.

Ref: https://www.ruby-lang.org/en/news/2020/12/25/ruby-3-0-0-released/

Not sure why this started to pop now, maybe something with the ruby
versions available in CircleCI. Anyway, it seems legit and we should do
it.