Remove automatic propagation of generators options
By removing the override of the `generate` method in this
generator, we stop propagating those custom arguments to
all the inner generators called by the installer.

For example, at the moment, the installer is running

	rails    generate rspec:install --auto-accept --auto-run-migrations

see [1], which doesn't make sense and might also have
unexpected behaviors if we (or any dependency) run a generator
during their installation, which accepts those arguments.

Instead of always propaging these option, we are explicitely
declaring them when needed.

[1]: https://github.com/solidusio/solidus_starter_frontend/issues/295

Co-authored-by: safafa <safaaballagh@nebulab.com>