Allow to disable track inventory for product without variants
We allow to disable track_inventory on the variant form, but
not on the product form. This makes it impossible to change
the value for sole products not having any variants, because
the value is stored on the master variant.

(cherry picked from commit 55920f30c53f877b8acdb3cf63c7e9dc88fc6cbc)