Add tailwindcss-rails as a dependency for solidus_admin
As with the other components, the development environment is configured
on the top Gemfile. We also need to add `solidus_admin` gem as a
dependency for `solidus`.

It's worth noticing that it's only a development dependency. The host
application won't need to use tailwind, as it'll just consume the
compiled CSS.

See https://github.com/rails/tailwindcss-rails